apps/agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.57.0
pydantic==2.10.3
python-multipart==0.0.20
asyncpg==0.30.0
redis==5.2.1
cryptography==44.0.0
python-dotenv==1.0.1
